Enable warnings as errors

+++ llarp/link/utp.cpp
@@ -125,7 +125,7 @@ namespace llarp
           llarp::LogDebug("utp_writev wrote=", s, " expect=", expect,
                           " to=", remoteAddr);
 
-          while(s > vecq.front().iov_len)
+          while(s > static_cast< ssize_t >(vecq.front().iov_len))
           {
             s -= vecq.front().iov_len;
             vecq.pop_front();
